refactor(kitchen pos) : PurgomalumClient 는 Menu나 Product에 대한 직접적인 종속성이 없어 port/adaptor 가 아닌 domain/sevice에 위치
