+++ src/main/java/kitchenpos/menu/domain/service/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.infrastructure.external;
+package kitchenpos.menu.domain.service;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
