+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -7,7 +7,7 @@ import kitchenpos.menu.domain.model.MenuProduct;
 import kitchenpos.menu.domain.repository.MenuRepository;
 import kitchenpos.menu.domain.model.Product;
 import kitchenpos.menu.domain.repository.ProductRepository;
-import kitchenpos.menu.infrastructure.external.PurgomalumClient;
+import kitchenpos.menu.domain.service.PurgomalumClient;
 import kitchenpos.menu.infrastructure.persistence.InMemoryMenuGroupRepository;
 import kitchenpos.menu.infrastructure.persistence.InMemoryMenuRepository;
 import kitchenpos.menu.infrastructure.persistence.InMemoryProductRepository;
